// src/interfaces/cartItem.interface.ts
export interface ICartItem {
  qtyOrdered: number;
  vid: string;
  sku: string;
  color: string;
  size: string;
  price: number;
  originalPrice: number;
  modelName: string;
  stock: number;
  images: string[];
  id: string;
  name: string;
  category: string;
  type: string;
  brand: string;
  currency: string;
  description: string;
  rating: number;
  productCategory: string;
  tags: string[];
  gender: string;
}